+++ iOS/Airbnb/Airbnb/Views/MapButton.swift
@@ -9,11 +9,7 @@
 import UIKit
 
 @IBDesignable
-final class MapButton: UIButton {
-    @IBInspectable var diameter: CGFloat = 50 {
-        didSet { configureDiameter() }
-    }
-
+final class MapButton: RoundButton {
     override init(frame: CGRect) {
         super.init(frame: frame)
         configureShadow(offset: CGSize(width: 5, height: 5), opacity: 3)
@@ -22,12 +18,6 @@ final class MapButton: UIButton {
     required init?(coder: NSCoder) {
         super.init(coder: coder)
         configureShadow(offset: CGSize(width: 5, height: 5), opacity: 3)
-    }
-    
-    private func configureDiameter() {
-        widthAnchor.constraint(equalToConstant: diameter).isActive = true
-        heightAnchor.constraint(equalToConstant: diameter).isActive = true
-        layer.cornerRadius = diameter / 2
     }
 }
